% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{create_project_folders}
\alias{create_project_folders}
\title{Create common project folders}
\usage{
create_project_folders()
}
\value{
Nothing
}
\description{
Create common project folders
}
\author{
Dave Fifield
}
