% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfexample.R
\name{dfexample}
\alias{dfexample}
\title{Example of fitting a step-wise mass function}
\usage{
dfexample(case = 1, seed = 1)
}
\arguments{
\item{case}{is an integer from 1 to 4, specifying the example. Choose 1 for an example of fitting a Schechter function mock data drawn from a Schechter function. Choose @ to fit a Schechter function to mock data drawn from a Schechter function in the presence of large-scale structure. Choose 3 to fit a quasi non-parametric mass function to mock data drawn from a modified Scheckter function. Choose 4 to fit a 2D analytic distribution function to mock data in the the mass-specific angular momentum plane.}

\item{seed}{A positive integer used as seed for the random number generator.}
}
\description{
This function generates n mock galaxies with masses and observing errors, drawn from a custom Schechter function and predefined effective volume function. It then uses \code{\link{dffit}} to recover the inpu Schechter function from the noisy data.
}
\examples{
# Run basic example
dfexample(1)

}
\seealso{
See \code{\link{dffit}}.
}
\author{
Danail Obreschkow
}
