% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfexample1.R
\name{dfexample1}
\alias{dfexample1}
\title{Example of using dftools}
\usage{
dfexample1(n = 10000, seed = 1, sigma = 0.5, p.true = c(-2, 10, -1.3),
  include.mrp = FALSE)
}
\arguments{
\item{n}{Number of gtalaxies}

\item{seed}{Random number seed}

\item{sigma}{Observational uncertainty, defined as standard-deviation in log10(Mass)}

\item{p.true}{Input parameters for the Schechter function}

\item{include.mrp}{If \code{TRUE}, an MRP function will also be fitted}
}
\description{
This function generates n mock galaxies with observing errors, drawn from a Schechter function with predefined parameters, using a predefined effective volume function. It then uses \code{\link{dffit}} to recover the input function.
}
\examples{
dfexample1()

}
\seealso{
See examples in \code{\link{dffit}}.
}
\author{
Danail Obreschkow
}
