% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmockdata.R
\name{dfmockdata}
\alias{dfmockdata}
\title{Example data}
\usage{
dfmockdata(n = 100, seed = 0, f = function(x, r) {     mass = 10^x    
  rmax = sqrt(mass)     f = pracma::erf((rmax - r)/(rmax + 1)) * 0.5 + 0.5 },
  dVdr.scale = function(r) {     r^2 }, gdf = function(x, p) dfmodel(x, p,
  type = "Schechter"), p = c(-2, 10, -1.3), sigma = 0.1, rmin = 0,
  rmax = 20, xmin = 0, xmax = 12, dx = 0.01)
}
\description{
This function produces an example of log-masses \code{x} with Gaussian uncertainties \code{x.err} and a selection \code{selection}. This selection is a list of an array with galaxy-by-galaxy effective volumes and a function to complete the effective volume beyond the observed mass range. The data is taken from a publication by Westmeier et al. (2017) and represents HI-Masses in the Sculptor structure.
}
\examples{
data <- dfdata()
print(data$x)
print(data$x.err)
print(data$selection[[1]])

# These data can then be used to fit a MF, e.g.
df = dffit(data$x, data$selection, data$x.err)

}
\seealso{
\code{\link{dffit}}
}
\author{
Danail Obreschkow
}
