% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmodel.R
\name{dfmodel}
\alias{dfmodel}
\title{Analytical distribution functions}
\usage{
dfmodel(x = NULL, p = NULL, output = "density", type = "Schechter")
}
\arguments{
\item{x}{Vector of log-masses, typically \code{x = log10(M/Msun)}.}

\item{p}{Parameters of the analytical function. See argument \code{type}.}

\item{output}{Specifies what the function is doing. \code{'density'} evaluates the MF at \code{x}, \code{'npara'} returns the number of parameters of the analytical function, \code{'initial'} returns a vector of typical parameters \code{p} that are used as default initial values when fitting the MF, \code{'equation'} returns a string with the equation of the MF.}

\item{type}{Kind of MF: \code{'Schechter'} for Schechter function; \code{'PL'} for a simple power law; \code{'MRP'} for a Murray-Robotham-Power function, a 4-parameter extension of the Schechter function.}
}
\description{
This function allows the user to call several pre-defined distribution functions (DFs). In the current implementation, they are all galaxy mass functions (i.e. a type of one-dimensional DFs), such as the Schechter function.
}
\examples{
# Evaluate and plot a Schechter function
x = seq(7,11,length=100)
mass = 10^x
parameters = c(-2,10,-1.5)
phi = dfmodel(x, parameters, type = 'Schechter')
plot(mass, phi, type='l', log='xy')

}
\author{
Danail Obreschkow
}
\seealso{
\code{\link{fit.df}}
}

