% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfplot.R
\name{dfplot}
\alias{dfplot}
\title{Display fitted distribution function}
\usage{
dfplot(df, xlab = "Observable x", ylab = expression("Distribution function"
  ~ phi), xlim = NULL, ylim = NULL, log = "y", col = "blue",
  show.uncertainties = TRUE, uncertainty.type = NULL)
}
\arguments{
\item{df}{List produced by \code{\link{dffit}}}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{xlim}{2-element vector with x-axis plotting limits}

\item{ylim}{2-element vector with y-axis plotting limits}

\item{show.uncertainties}{If \code{TRUE}, uncertainties are displayed around the best fit model.}

\item{uncertainty.type}{\code{1}: plot Gaussian 1-sigma uncertanties propagated from the Hessian matrix of the likelihood. \code{2}: plot 68 percentile region (from 16 to 84 percent). \code{3} plot 68 (16 to 84) and 95 (2 to 98) percentile regions.}
}
\description{
This function displays the distribution function fitted using \code{\link{dffit}}.
}
\author{
Danail Obreschkow
}
\seealso{
See examples in \code{\link{dffit}}.
}

