% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfplot2.R
\name{dfplot2}
\alias{dfplot2}
\title{Display fitted distribution function}
\usage{
dfplot2(survey, xlab = "Observable x1", ylab = "Observable x2",
  xlim = NULL, ylim = NULL, p.ref = NULL, xpower10 = TRUE,
  ypower10 = TRUE, show.data = TRUE, show.data.err = TRUE,
  show.phi = TRUE, show.phi.contours = TRUE, show.sc.contours = TRUE,
  contour.levels = c(0.5, 0.2, 0.1), gamma = 0.2, col.data = "black",
  cex.data = 0.5, pch.data = 20, col.data.err = "#00000030",
  lwd.data.err = 1, lty.data.err = 1, col.phi = "blue",
  col.phi.contours = col.phi, lwd.phi.contours = 1, lty.phi.contours = 2,
  col.sc.contours = "purple", lwd.sc.contours = 1, lty.sc.contours = 1,
  col.ref.contours = "red", lwd.ref.contours = 1, lty.ref.contours = 2,
  margins = c(5.1, 4.1, 4.1, 2.1))
}
\arguments{
\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{xlim}{2-element vector with x-axis plotting limits}

\item{ylim}{2-element vector with y-axis plotting limits}

\item{xpower10}{If \code{TRUE}, the model argument x is elevated to the power of 10 in the plots.}

\item{ypower10}{If \code{TRUE}, the model argument x is elevated to the power of 10 in the plots.}

\item{show.data}{If \code{TRUE}, binned data points are displayed.}

\item{show.data.err}{If \code{TRUE}, the bias corrected MLE is shown instead of the native ML parameters.}

\item{margins}{Margins (bottom,left,top,right)}

\item{df}{List produced by \code{\link{dffit}}}

\item{col}{Color of fit (see \code{\link{plot}})}

\item{cex}{Line width of fit (see \code{\link{plot}})}

\item{pch}{Line type of fit (see \code{\link{plot}})}
}
\value{
Returns the input list \code{df} with the additional sub-list \code{survey$bin} that contains the binned data.
}
\description{
This function displays a one-dimensional distribution function fitted using \code{\link{dffit}}.
}
\seealso{
For optimized plotting of galaxy mass functions, use the derived function \code{\link{mfplot}}. See examples in \code{\link{dffit}}.
}
\author{
Danail Obreschkow
}
