% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfplotcov.R
\name{dfplotcov}
\alias{dfplotcov}
\title{Plot parameter covariance matrix}
\usage{
dfplotcov(data, name = NULL, title = "", nstd = 10, lower = TRUE,
  upper = FALSE, margins = c(4, 4, 0.5, 0.5), col = c("black", "blue",
  "red", "green", "orange"), lwd = 1, lty = 1, cex = 1.2, pch = 20,
  cloud.alpha = 0.1, cloud.nmax = NULL, hist.alpha = 0.4,
  show.histogram = TRUE, show.cloud = TRUE, show.gaussian = TRUE,
  show.expectation = TRUE, show.ellipse.68 = TRUE, show.ellipse.95 = TRUE,
  text.size.labels = 1.1, text.size.numbers = 0.8,
  text.offset.labels = c(0, 0), text.offset.numbers = c(0, 0),
  text.format.numbers = "\%4.1f")
}
\arguments{
\item{data}{List of objects to be plotted. Each object in this list must be one of the following four: (1) a vector of parameters; (2) a list of two objects, a vector of parameters and an associated covariance matrix; (3) a matrix, where each row represents a set of model parameters, to be displayed as a point of a cloud; (4) an output list produced by \code{\link{dffit}}. The plot will be centered on the parameters or average parameters of the first object in the data list.}

\item{name}{Optional list of parameter names}

\item{title}{Optional plot title}

\item{nstd}{Width of the plots in multiples of the standard deviations of the model, i.e. the square roots of the diagonal elements of \code{covariance}.}

\item{lower}{Logical flag indicating whether the lower triangle is shown.}

\item{upper}{Logical flag indicating whether the lower triangle is shown.}

\item{margins}{Plot margins (bottom,left,top,right)}

\item{col}{Vector of colors of each object in the data-list.}

\item{lwd}{Vector of line width of each object in the data-list.}

\item{lty}{Vector of line types of each object in the data-list.}

\item{cex}{Vector of point sizes for the central parameters each object in the data-list.}

\item{pch}{Vector of point type for the central parameters each object in the data-list.}

\item{text.size.labels}{Text size of parameter names}

\item{text.size.numbers}{Text size of numbers}

\item{text.offset.labels}{2-element vector to adjust the position of the vertical and horizontal parameter names}

\item{text.offset.numbers}{2-element vector to adjust the position of the vertical and horizontal numbers}

\item{text.format.numbers}{String specifying the floating point format of the numbers. (see \code{\link{sprintf}})}
}
\description{
This function displays the parameter covariances  of the distribution function parameters (e.g. the mass function parameters) fitted using \code{\link{dffit}}.
}
\seealso{
See examples in \code{\link{dffit}}.
}
\author{
Danail Obreschkow
}
