% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfplotveff.R
\name{dfplotveff}
\alias{dfplotveff}
\title{Plot effective volume}
\usage{
dfplotveff(bundle, xlab = "Observable x",
  ylab = expression("Effective volume"))
}
\arguments{
\item{xlab}{Label on x-axis (logarithmic mass or luminosity scale).}

\item{ylab}{Label on y-axis (volume scale).}

\item{df}{List produced by \code{\link{dffit}}}
}
\description{
This function plots the function \code{Veff(x)} used for the mass function fit, as stored in the sublist \code{fit} of the output produced by \code{\link{dffit}}.
}
\seealso{
See examples in \code{\link{dffit}}.
}
\author{
Danail Obreschkow
}
