% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfplotveff2.R
\name{dfplotveff2}
\alias{dfplotveff2}
\title{Plot effective volume in 2D}
\usage{
dfplotveff2(survey, xlab = "Observable x1", ylab = "Observable x2",
  zlab = "Veff")
}
\arguments{
\item{survey}{List produced by \code{\link{dffit}}}

\item{xlab}{Label on x-axis, associated with the first observable (\code{x[1,]}) in the 2D data (\code{survey$data$x}).}

\item{ylab}{Label on y-axis, associated with the first observable (\code{x[2,]}) in the 2D data (\code{survey$data$x}).}

\item{zlab}{Label on z-axis, displaying the GDF as a function of x and y.}
}
\description{
This routine plots the effective survey volume function \code{Veff(x)} assumed when fitting a two-dimensional generative distribution function (GDF). Note that this function \code{Veff(x)} is stored as \code{survey$selection$veff} when fitting a GDF using \code{survey=dffit(...)}.
}
\seealso{
See examples in \code{\link{dffit}}. To display \code{Veff(x)} of one-dimensional distribution functions use \code{\link{dfplotveff}}.
}
\author{
Danail Obreschkow
}
