% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfposteriors.R
\name{dfposteriors}
\alias{dfposteriors}
\title{Evaluate posterior masses}
\usage{
dfposteriors(df)
}
\arguments{
\item{df}{List produced by \code{\link{dffit}}}
}
\value{
Returns a structured list, which, in addition to the input argument, also contains the sublist \code{posteriors}. Note that the posteriors can be visualized using \code{\link{dfplot}} with the argument \code{bin.use.posteriors = TRUE}.
}
\description{
Given a fitted df, this function computes posterior masses.
}
\seealso{
See examples in \code{\link{dffit}}.
}
\author{
Danail Obreschkow
}
