% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfssmodel.R
\name{dfssmodel}
\alias{dfssmodel}
\title{Make step-wise linearmass function}
\usage{
dfssmodel(xpoints, method = "linear", p.Schechter = dfmodel(output =
  "initial"))
}
\arguments{
\item{xpoints}{Vector of log-masses used as interpolation points.}

\item{p}{Schechter function parameters used to determine the initial values of the step-wise paraemters.}
}
\value{
\code{dfssmodel} returns a list with two components \code{gdf} and {p.initial}. The first component is a function of log-masses \code{x} and a parameter vector \code{p}. The second component is the initial parameter-vector producing a step-wise mass function that closely matches the Schechter function specified by \code{p}.
}
\description{
This function converts a Schechter function into a step-wise linear mass function of an arbitrary number of bins, to fit a quasi non-parametric mass functions.
}
\examples{
# Basic example of a stepwise linear MF
xpoints = seq(6.5,12,0.5)
model = dfssmodel(xpoints)
plot(10^xpoints,10^model$p.initial,log='xy',ylim=c(1e-5,1),pch=20)
lines(10^seq(4,12,0.01),model$gdf(seq(4,12,0.01),model$p.initial))

# Example of fitting a stepwise linear MF
dfexample4()

}
\seealso{
\code{\link{dffit}}
}
\author{
Danail Obreschkow
}
