% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAccuracy.R
\name{getAccuracy}
\alias{getAccuracy}
\title{Calculates the accuracy}
\usage{
getAcuracy(score, response, 0.5)
}
\arguments{
\item{score}{prediction scores or probabilities to evaluate.}

\item{response}{Boolean vector of real values.}

\item{unbral}{cut point for scores, check maxPrecision function.}
}
\value{
Returns the accuracy
}
\description{
Calculates the accuracy of a prediction
}
\author{
Daniel Fischer
}
