% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConfusion.R
\name{getConfusion}
\alias{getConfusion}
\title{Calculates the confussion matrix}
\usage{
getConfusion(score, response, 0.5)
}
\arguments{
\item{score}{prediction scores or probabilities to evaluate.}

\item{response}{Boolean vector of real values.}

\item{asMatric}{Should the function return a matrix object? otherwise will be a dataframe.}

\item{unbral}{should confussion matrix be retourned as percentages.}
}
\value{
RReturns the conffusin matrix
}
\description{
Calculates the confussion matrix
}
\author{
Daniel Fischer
}
