% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPrecision.R
\name{getPrecision}
\alias{getPrecision}
\title{Calculates the Precission}
\usage{
getPrecision(score, response, 0.5)
}
\arguments{
\item{score}{prediction scores or probabilities to evaluate.}

\item{response}{Boolean vector of real values.}

\item{unbral}{cut point for scores, check maxPrecision function.}
}
\value{
Returns the precission
}
\description{
Calculates the Precission
}
\author{
Daniel Fischer
}
