% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRecall.R
\name{getRecall}
\alias{getRecall}
\title{Calculates the Recall}
\usage{
getRecall(score, response, 0.5)
}
\arguments{
\item{score}{prediction scores or probabilities to evaluate.}

\item{response}{Boolean vector of real values.}

\item{unbral}{cut point for scores, check maxPrecision function.}
}
\value{
Returns the recall
}
\description{
Calculates the Recall
}
\author{
Daniel Fischer
}
