% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{install_standard_packages}
\alias{install_standard_packages}
\title{Install standard set of packages}
\usage{
install_standard_packages()
}
\value{
noting
}
\description{
Install standard set of packages
}
\author{
Dave Fifield
}
