% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is_even}
\alias{is_even}
\title{Tell if a number is even}
\usage{
is_even(num)
}
\arguments{
\item{num}{Number to check}
}
\value{
\code{TRUE} if \code{num} is even, \code{FALSE} otherwise.
}
\description{
Tell if a number is even
}
\author{
Dave Fifield
}
