% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is_odd}
\alias{is_odd}
\title{Tell if a number is odd.}
\usage{
is_odd(num)
}
\arguments{
\item{num}{Number to check}
}
\value{
\code{TRUE} if \code{num} is odd, \code{FALSE} otherwise.
}
\description{
Tell if a number is odd.
}
\author{
Dave Fifield
}
