% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeScoreDistribution.R
\name{makeScoreDistribution}
\alias{makeScoreDistribution}
\title{Provides Score Distribution interpoles}
\usage{
makeScoreDistribution(score, response, method = "approx", groups = 100, ...)
}
\arguments{
\item{score}{a numeric vector with the scores from the prodictive model}

\item{response}{a boolean (TRUE or FALSE) vector with the real values from the dataset}

\item{method}{method used to interpolate scores between groups, approx for approxfun or spline for splinefun}

\item{groups}{number of percentiles, default 100}

\item{...}{passed to aproximation method}
}
\value{
data.frame with the values of the Lift curve.
score = runif(1000)
response = (score + rnorm(1000,0,0.1)) > 0.5
fn = makeScoreDistribution(score,response,methond = 'approx')
fn(0.5)
}
\description{
Calculates the lift curve based on a score(numeric 0-1) and responce(TRUE/FALSE) vector
}
\author{
Daniel Fischer
}
