% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxAccuracy.R
\name{maxAccuracy}
\alias{maxAccuracy}
\title{Maximizes the accuracy}
\usage{
getAcuracy(score, response, 0.5)
}
\arguments{
\item{score}{prediction scores or probabilities to evaluate.}

\item{response}{Boolean vector of real values.}
}
\value{
returns a list with the optimization results, where metrics element has the report.
}
\description{
Finds th optimal cut point in order to maximize the accuracy
}
\author{
Daniel Fischer
}
