% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfdata.R
\name{mfdata}
\alias{mfdata}
\title{Example data}
\usage{
mfdata()
}
\description{
This function produces an example of log-masses \code{x} with Gaussian uncertainties \code{x.err} and a selection \code{selection}. This selection is a list of an array with galaxy-by-galaxy effective volumes and a function to complete the effective volume beyond the observed mass range. The data is taken from a publication by Westmeier et al. (2017) and represents HI-Masses in the Sculptor structure.
}
\examples{
data <- mfdata()
print(data$x)
print(data$x.err)
print(data$selection[[1]])

# These data can then be used to fit a MF, e.g.
df = dffit(data$x, data$veff, data$x.err)

}
\author{
Danail Obreschkow
}
\seealso{
\code{\link{fit.df}}
}

