% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfexample.R
\name{mfexample}
\alias{mfexample}
\title{Example of using mftools}
\usage{
mfexample(n = 25000, seed = 0, sigma = 0.5, p.true = c(-2, 9, -1.3),
  include.mrp = FALSE)
}
\arguments{
\item{n}{Number of gtalaxies}
}
\description{
This function generates n mock galaxies with observing errors, drawn from a predefined mass function and effective observing volume. It then uses \code{\link{fit.df}} to recover the input function
}
\examples{
example()

}
\author{
Danail Obreschkow
}
\seealso{
See examples in \code{\link{dffit}}.
}

