% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{namesort}
\alias{namesort}
\title{Return sorted names of an object}
\usage{
namesort(obj)
}
\arguments{
\item{obj}{Object to get names of}
}
\value{
Sorted vector of object names.
}
\description{
Return sorted names of an object
}
\author{
Dave Fifield
}
