% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{setup_project}
\alias{setup_project}
\title{Setup a new project}
\usage{
setup_project(create.folders = TRUE, install.pkgs = TRUE, use.renv = TRUE)
}
\arguments{
\item{create.folders}{[Boolean, default TRUE] Should folder hierarchy be
created. The following folders will be created under the project root:
"R", "Results", "GIS/Shapefiles", "GIS/Maps", "Data/Raw", "Data/Generated",
"Literature", "Tests"}

\item{install.pkgs}{[Boolean, default TRUE] Should standard set of packages
be installed: The following packages will be installed: "here", "tidyverse",
"usethis", "devtools".}

\item{use.renv}{[Boolean, default TRUE] Indicates whether \code{renv} is in
use with this project. If \code{TRUE} and \code{install.packages}
is \code{TRUE}, then \link{renv::snapshot()} will be called after installing
packages.}
}
\value{
Nothing
}
\description{
Setup a new project
}
\author{
Dave Fifield
}
