% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlInsert.R
\name{sqlInsert}
\alias{sqlInsert}
\title{saves dataframe on SQL row by row}
\usage{
sqlWriteTable(server_name, data, table, ...)
}
\arguments{
\item{server_name}{character, name of the DB server from sqlServers list.}

\item{data}{a data frame or coercible to data frame}

\item{table}{character, name of the table to write in DB.}
}
\value{
returns booleans as it is specificated for dbWriteTable().
}
\description{
saves dataframe on SQL row by row
It calls dbWriteTable().
}
\details{
It ends the connection inmediately after getting the results. sqlServers is
a list built-in sqlGetConn().
}
\examples{
sqlInsert(local,head(iris),"iris")
sqlInsert(local,head(mtcars),"MTCARS")
}
\seealso{
"sqlServerConn" ocumentation in dftools and "dbWriteTable()"
from DBI for more details.
}
