% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfexample2.R
\name{dfexample2}
\alias{dfexample2}
\title{Example of using mftools}
\usage{
dfexample2(p.true = c(-2, 10, -1.3, 2/3, 7, 0.3), seed = 3,
  sigma.x = 0.25, sigma.y = 0.1, corr.xy = 0.7)
}
\arguments{
\item{n}{Number of gtalaxies}
}
\description{
This function generates n mock galaxies with observing errors, drawn from a predefined mass function and effective observing volume. It then uses \code{\link{dffit}} to recover the input function
}
\examples{
dfexample2()

}
\seealso{
See examples in \code{\link{dffit}}.
}
\author{
Danail Obreschkow
}
