% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfexample2.R
\name{dfexample2}
\alias{dfexample2}
\title{Example of fitting a galaxy mass function in the presence of large-scale structure}
\usage{
dfexample2(seed = 3, sigma = 0.3)
}
\arguments{
\item{seed}{A positive integer used as seed for the random number generator.}

\item{sigma}{Gaussian observing error in log10(mass).}
}
\description{
This function generates n mock galaxies with masses, mass errors and distances. The mock sample is drawn from a custom Schechter function, predefined selection function and predefined large-scale structure density \code{f(r)}. The function then runs \code{\link{dffit}} to recover the inpu Schechter function from these noisy and LSS-biased data.
}
\examples{
dfexample2()

}
\seealso{
See examples in \code{\link{dffit}}.
}
\author{
Danail Obreschkow
}
