% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmockdata.R
\name{dfmockdata}
\alias{dfmockdata}
\title{Generate mock data}
\usage{
dfmockdata(n = NULL, seed = 2, f = function(x, r) pracma::erf((1 - 100 *
  r/sqrt(10^x)) * 20) * 0.5 + 0.5, dVdr = function(r) 0.02439568 * r^2,
  gdf = function(x, p) dfmodel(x, p, type = "Schechter"), p = c(-2, 10,
  -1.3), g = function(r) 1, sigma = 0, rmin = 0, rmax = 100, xmin = 2,
  xmax = 12, shot.noise = FALSE, verbose = FALSE)
}
\arguments{
\item{n}{Number of objects (galaxies) to be generated. If \code{n=NULL}, the number is determined from the mass function (\code{gdf}) and the selection criteria (specified by \code{f} and \code{dVdr}). Otherwise, the survey volume (specified by the derivative \code{dVdr}) is automatically multiplied by the scaling factor required to obtain the requested number of objects \code{n}.}

\item{seed}{An interger number used as seed for the random number generator. If you wish to generate different realizations, with the same survey specifications, it suffices to vary this number.}

\item{f}{is the selection function \code{f(x,r)}, giving the ratio between the expected number of detected galaxies and true galaxies of log-mass \code{x} and comoving distance \code{r}. Normally this function is bound between 0 and 1. It takes the value 1 at distances, where objects of mass \code{x} are easily detected, and 0 at distances, where such objects are impossible to detect. A rapid, continuous drop from 1 to 0 normally occurs at the limting distance \code{rmax}, at which a galaxy of log-mass \code{x} can be picked up. \code{f(x,r)} can never by smaller than 0, but values larger than 1 are conceivable, if there is a large number of false positive detections in the survey.}

\item{dVdr}{is the function \code{dVdr(r)}, spedifying the derivative of the survey volume \code{V(r)} as a function of comoving distance \code{r}. This survey volume is simply the total observed volume, irrespective of the detection probability, which is already specified by the function \code{f}. Normally, the survey volume is given by \code{V(r)=Omega*r^3/3}, where \code{Omega} is the solid angle of the survey. Hence, the derivative is \code{dVdr(r)=Omega*r^2}. The default uses \code{Omega=2.439568e-2} [sterradians], such that the expected number of galaxies is exactly 1000.}

\item{gdf}{is the 'generative distribution function', i.e. the underlying mass function, from which the galaxies are drawn. This function is a function of log-mass \code{x} and model parameters \code{p}. It returns the expected number of galaxies per unit of cosmic volume \code{V} and log-mass \code{x}. The default is a Schechter function.}

\item{p}{model parameters for the \code{gdf}.}

\item{g}{function of distance \code{r} descibing the number-density variation of galaxies due to cosmic large-scale structure (LSS). Explicitly, \code{g(r)>0} is the number-density at \code{r}, relative to the number-density without LSS. Values between 0 and 1 are underdense regions, values larger than 1 are overdense regions. In the absence of LSS, \code{g(r)=1}. Note that g is automatically rescaled, such that its average value in the survey volume is 1.}

\item{sigma}{Gaussian observing errors in log-mass \code{x}, which are automatically added to the survey.}

\item{rmin, rmax}{Minimum and maximum distance of the survey. Outside these limits the function \code{f(x,r)} will automatically be assumed to be 0.}

\item{xmin, xmax}{Minimum and maximum log-mass in the survey. For optimal performance, specify these boubdaries in such a way that they certainly contain all sources generated by the survey, but don't span a much larger range.}

\item{shot.noise}{Logical flag. If set to \code{TRUE}, the number of galaxies in the survey can differ from the expected number, following a Poisson distribution.}

\item{verbose}{Logical flag. If set to \code{TRUE}, some information will be displayed in the console while generating the mock survey.}
}
\value{
\code{dfmockdata} returns a list of arrays and scalars:
\item{x}{Array of observed log-mass.}
\item{x.err}{Gaussian uncertainties on x.}
\item{x.true}{Array of true log-masses, i.e. the values of \code{x} before they were perturbed by random uncertainties \code{x.err}.}
\item{r}{Array of comoving distances.}
\item{f}{Selection function provided as input argument.}
\item{g}{Cosmic LSS function provided as input argument.}
\item{dVdr}{Derivative of survey volume provided as input argument, but rescaled to the requested number of galaxies \code{n}.}
\item{veff}{Function returning the effective volume as a function of log-mass \code{x}.}
\item{veff.values}{Array of effective volumes for each galaxy.}
\item{scd}{Function returning the expected source count density as a function of log-mass \code{x}.}
\item{rmin,rmax}{Range of comoving distances \code{r}, spanned by the survey. Same as input arguments.}
\item{xmin,xmax}{Range of log-masses \code{x} provided as input argument. This range is generally larger than the range spanned by the values of \code{x} and is meant to span the maximally conceivable range of \code{x} given the survey specifications.}
\item{rescaling.factor}{Value of rescaling factor applied to the cosmic volume to match the requested number of galaxies \code{n}.}
}
\description{
This function produces a mock survey with observed log-masses \code{x.obs} with Gaussian uncertainties \code{x.err} and distances \code{r}, using a custom mass function (MF) and selection function.
}
\examples{
dat = dfmockdata(sigma = 0.3)
plot(dat$r,dat$x,col='blue')
points(dat$r,dat$x.true,pch=20)
x = seq(5,11,0.01)
lines(1e-2*sqrt(10^x),x,col='red')

# These data can then be used to fit a MF in several ways. For instance,
# assuming that the effective volume function Veff(x) is known:
selection = dat$veff
survey = dffit(dat$x, selection, dat$x.err)

# or assuming that Veff is known only on a galaxy-by-galaxy basis
selection = dat$veff.values
dffit(dat$x, selection, dat$x.err)

# or assuming that Veff is known on a galaxy-by-balaxy basis, but approximate analytically
# outside the range of observed galaxy masses
selection = list(dat$veff.values, dat$veff)
dffit(dat$x, selection, dat$x.err)

# or assuming that the full selection function f(x,r) and the observing volume
# derivative dVdr(r) are known
selection = list(dat$f, dat$dVdr, dat$rmin,dat$rmax)
dffit(dat$x, selection, dat$x.err)

}
\seealso{
\code{\link{dffit}}
}
\author{
Danail Obreschkow
}
