% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfplot.R
\name{dfplot}
\alias{dfplot}
\title{Display fitted generative distribution function}
\usage{
dfplot(
  survey,
  xlab = "Observable x",
  ylab = expression("Generative distribution function" ~ phi),
  ylab.histogram = "Counts",
  xlim = NULL,
  ylim = NULL,
  log = "y",
  p = NULL,
  veff = NULL,
  xpower10 = FALSE,
  show.input.data = TRUE,
  show.posterior.data = TRUE,
  show.data.histogram = FALSE,
  show.uncertainties = TRUE,
  uncertainty.type = 1,
  show.bias.correction = FALSE,
  add = FALSE,
  scd = NULL,
  nbins = NULL,
  bin.xmin = NULL,
  bin.xmax = NULL,
  clip.xmin = NULL,
  clip.xmax = NULL,
  clip.ymin = NULL,
  clip.ymax = NULL,
  col.fit = "blue",
  lwd.fit = 2,
  lty.fit = 1,
  col.data.input = "grey",
  cex.data.input = 1,
  lwd.data.input = 1,
  col.data.posterior = "blue",
  cex.data.posterior = 1,
  lwd.data.posterior = 1,
  col.hist = "grey",
  col.ref = "black",
  lwd.ref = 1,
  lty.ref = 2,
  col.veff = "#666666",
  lwd.veff = 1.5,
  lty.veff = 3,
  col.scd = "black",
  lwd.scd = 1,
  lty.scd = 1,
  div.line = TRUE,
  axes = TRUE,
  margins = c(5.1, 4.1, 4.1, 2.1)
)
}
\arguments{
\item{survey}{List produced by \code{\link{dffit}}}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{ylab.histogram}{y-axis label for histogram}

\item{xlim}{2-element vector with x-axis plotting limits}

\item{ylim}{2-element vector with y-axis plotting limits}

\item{log}{String specifying the log-axes as in \code{\link{plot}}.}

\item{p}{Parameters of a reference distribution function do be over-plotted to the fitted function. Choose \code{NULL} to avoid plotting a reference function.}

\item{veff}{Function of x to be plotted over the histogram, if \code{show.data.histogram = TRUE}.}

\item{xpower10}{If \code{TRUE}, the model argument x is elevated to the power of 10 in the plots.}

\item{show.input.data}{If \code{TRUE}, the input data is shown in bins. Each bin values is simply the sum 1/Veff(x) of the observed x-values in this bin.}

\item{show.posterior.data}{If \code{TRUE}, the posterior data, constructucted from all the individual posterior PDFs of the observed data, are snown in bins. Note that posterior data only exists of the fitted data is uncertain, e.g. if the argument \code{x.err} in \code{\link{dffit}} was non-zero.}

\item{show.data.histogram}{If \code{TRUE}, a histogram of source counts, based on the input data, is displayed in a bottom panel. Choose \code{NA}, to leave space for the histogram without drawing it.}

\item{show.uncertainties}{If \code{TRUE}, uncertainties are displayed around the best fit model.}

\item{uncertainty.type}{\code{1}: plot Gaussian 1-sigma uncertanties propagated from the Hessian matrix of the likelihood. \code{2}: plot 68 percentile region (from 16 to 84 percent). \code{3} plot 68 (16 to 84) and 95 (2 to 98) percentile regions.}

\item{show.bias.correction}{If \code{TRUE}, the bias corrected MLE is shown instead of the native ML parameters.}

\item{add}{If \code{TRUE}, the lines are overplotted on the currently open plot.}

\item{scd}{is an optional N-vector or scalar function specifying a source count density to be plotted over the histogram.}

\item{nbins}{Number of bins to be plotted; must be larger than 0. Choose \code{nbins=NULL} (default) to determine the number of bins automatically.}

\item{bin.xmin}{Left edge of first bin}

\item{bin.xmax}{Right edge of last bin}

\item{clip.xmin}{Minimum x values to be plotted in models}

\item{clip.xmax}{Maximum x values to be plotted in models}

\item{clip.ymin}{Minimum y values to be plotted in models}

\item{clip.ymax}{Maximum y values to be plotted in models}

\item{col.fit}{Color of fit (see \code{\link{plot}})}

\item{lwd.fit}{Line width of fit (see \code{\link{plot}})}

\item{lty.fit}{Line type of fit (see \code{\link{plot}})}

\item{col.data.input}{Color of binned input data}

\item{cex.data.input}{Size of binned input data}

\item{lwd.data.input}{Line width of binned input data}

\item{col.data.posterior}{Color of binned posterior data}

\item{cex.data.posterior}{Size of binned posterior data}

\item{lwd.data.posterior}{Line width of binned posterior data}

\item{col.hist}{Color of source count histogram}

\item{col.ref}{Color of reference distribution function}

\item{lwd.ref}{Line width of reference distribution function}

\item{lty.ref}{Line type of reference distribution function}

\item{col.veff}{Color of reference function \code{veff}}

\item{lwd.veff}{Line width of reference function \code{veff}}

\item{lty.veff}{Line type of reference function \code{veff}}

\item{col.scd}{Color of source count density function \code{scd}}

\item{lwd.scd}{Line width of source count density function \code{scd}}

\item{lty.scd}{Line type of source count density function \code{scd}}

\item{div.line}{Logical flag specifying if the division line between upper and lower p}

\item{axes}{Logical flag specifying whether to plot axes}

\item{margins}{Margins (bottom,left,top,right)}
}
\value{
Returns the input list \code{survey} with the additional sub-list \code{survey$bin} that contains the binned data.
}
\description{
This function displays a one-dimensional generative distribution function fitted using \code{\link{dffit}}.
}
\seealso{
For optimized plotting of galaxy mass functions, use the derived function \code{\link{mfplot}}. As an example run \code{dfexample(1)}. See examples in \code{\link{dffit}}.
}
\author{
Danail Obreschkow
}
