% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfplot2.R
\name{dfplot2}
\alias{dfplot2}
\title{Display fitted distribution function}
\usage{
dfplot2(survey, xlab = "Observable x1", ylab = "Observable x2",
  xlim = NULL, ylim = NULL, p.ref = NULL, xpower10 = TRUE,
  ypower10 = TRUE, show.data = TRUE, show.data.err = TRUE,
  show.gdf = TRUE, show.gdf.contours = TRUE, show.scd.contours = TRUE,
  contour.levels = c(0.5, 0.2, 0.1), col.data = "black", cex.data = 0.5,
  pch.data = 20, col.data.err = "#00000030", lwd.data.err = 1,
  lty.data.err = 1, col.gdf = "blue", col.gdf.contours = col.gdf,
  lwd.gdf.contours = 1, lty.gdf.contours = 2, gamma.gdf = 0.2,
  col.scd.contours = "purple", lwd.scd.contours = 1, lty.scd.contours = 1,
  col.ref.contours = "red", lwd.ref.contours = 1, lty.ref.contours = 2,
  margins = c(5.1, 4.1, 4.1, 2.1))
}
\arguments{
\item{survey}{List produced by \code{\link{dffit}}}

\item{xlab, ylab}{axis labels}

\item{xlim, ylim}{2-element vectors with plotting ranges}

\item{p.ref}{Optional vector of parameters of a reference mdeol to be displayed.}

\item{xpower10}{If \code{TRUE}, the 1st model argument is elevated to the power of 10 in the plots.}

\item{ypower10}{If \code{TRUE}, the 2nd model argument is elevated to the power of 10 in the plots.}

\item{show.data}{Logical flag to show/hide the observed data.}

\item{show.data.err}{Logical flag to show/hide the measurement errors of the}

\item{show.gdf}{Logical flag to show/hide the best fitting model of the generative distribution function as gradual shading.}

\item{show.gdf.contours}{Logical flag to show/hide the best fitting model of the generative distribution function as iso-contours.}

\item{show.scd.contours}{Logical flag to show/hide the source count density (for data with no measurement errors) predicted by the best fitting model.}

\item{contour.levels}{Vector specifying the fractions of the integrated densities contained inside the contours of the gdf and scd.}

\item{col.data}{Color (see \code{\link{plot}}) of data points.}

\item{cex.data}{Size (see \code{\link{plot}}) of data points.}

\item{pch.data}{Point type (see \code{\link{plot}}) of data points.}

\item{col.data.err}{Color (see \code{\link{plot}}) of error ellipses of data.}

\item{lwd.data.err}{Line width (see \code{\link{plot}}) of error ellipses of data.}

\item{lty.data.err}{Line type (see \code{\link{plot}}) of error ellipses of data.}

\item{col.gdf}{Color (see \code{\link{plot}}) of best fitting generative distribution function.}

\item{gamma.gdf}{Scalar value >0 specifying the brightness-scale of the best fitting generative distribution function.}

\item{margins}{Margins (bottom,left,top,right)}

\item{col.gdf}{Color (see \code{\link{plot}}) of best fitting generative distribution function contours.}

\item{lwd.gdf}{Line width (see \code{\link{plot}}) of best fitting generative distribution function contours.}

\item{lty.gdf}{Line type (see \code{\link{plot}}) of best fitting generative distribution function contours.}

\item{col.gdf}{Color (see \code{\link{plot}}) of the source count density contours predicted by the best fitting model.}

\item{lwd.gdf}{Line width (see \code{\link{plot}}) of the source count density contours predicted by the best fitting model.}

\item{lty.gdf}{Line type (see \code{\link{plot}}) of the source count density contours predicted by the best fitting model.}

\item{col.ref}{Color (see \code{\link{plot}}) of the reference model contours specified by \code{p.ref}.}

\item{lwd.ref}{Line width (see \code{\link{plot}}) of the reference model contours specified by \code{p.ref}.}

\item{lty.ref}{Line type (see \code{\link{plot}}) of the reference model contours specified by \code{p.ref}.}
}
\value{
Returns the input list \code{survey}.
}
\description{
This function displays a one-dimensional distribution function fitted using \code{\link{dffit}}.
}
\seealso{
For an full example of \code{dfplot2} run \code{\link{dfexample2()}}.
}
\author{
Danail Obreschkow
}
