% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfplotcov.R
\name{dfplotcov}
\alias{dfplotcov}
\title{Plot parameter covariance matrix}
\usage{
dfplotcov(survey = NULL, expectation = NULL, covariance = NULL,
  expectation2 = NULL, covariance2 = NULL, chain = NULL,
  reference = NULL, name = NULL, title = "", nstd = 10, lower = TRUE,
  upper = FALSE, margins = c(4, 4, 0.5, 0.5), model.col = "blue",
  model.lwd = 2, model.lty = 1, model.cex = 1, model2.col = "black",
  model2.lwd = 2, model2.lty = 1, model2.cex = 1, chain.col = "purple",
  chain.lwd = 2, chain.lty = 1, chain.cex = 1, chain.point.cex = 0.1,
  chain.point.alpha = 0.2, reference.col = "black", reference.lwd = 1,
  reference.lty = 2, reference.cex = 1, text.size.labels = 1.1,
  text.size.numbers = 0.8)
}
\arguments{
\item{survey}{List produced by \code{\link{dffit}}}

\item{expectation}{Vector of expected parameters}

\item{covariance}{Matrix of parameter covariances}

\item{expectation2}{Vector of expected parameters of an optional 2nd model}

\item{covariance2}{Matrix of parameter covariances of an optional 2nd model}

\item{chain}{Matrix, where each row represents a set of model parameters, to be displayed as a point of a cloud.}

\item{reference}{Reference parameters to be plotted as points and crosses.}

\item{name}{Optional list of parameter names}

\item{title}{Optional plot title}

\item{nstd}{Width of the plots in multiples of the standard deviations of the model, i.e. the square roots of the diagonal elements of \code{covariance}.}

\item{lower}{Logical flag indicating whether the lower triangle is shown.}

\item{upper}{Logical flag indicating whether the lower triangle is shown.}

\item{margins}{Plot margins (bottom,left,top,right)}

\item{model.col}{Model color}

\item{model.lwd}{Model line width}

\item{model.lty}{Model line type}

\item{model.cex}{Model point size}

\item{model2.col}{Second model color}

\item{model2.lwd}{Second model line width}

\item{model2.lty}{Second model line type}

\item{model2.cex}{Second model point size}

\item{chain.col}{Color of point cloud}

\item{chain.lwd}{Line width of point cloud ellipses}

\item{chain.lty}{Line type of point cloud ellipses}

\item{chain.cex}{Point size at the center of point cloud}

\item{chain.point.cex}{Point size of cloud points}

\item{chain.point.alpha}{Transparency of cloud points}

\item{reference.col}{Color of reference point}

\item{reference.lwd}{Line width of reference lines}

\item{reference.lty}{Line type of reference lines}

\item{reference.cex}{Size of reference point}

\item{text.size.labels}{Text size of parameter names}

\item{text.size.numbers}{Text size of numbers}
}
\description{
This function displays the parameter covariances  of the distribution function parameters (e.g. the mass function parameters) fitted using \code{\link{dffit}}.
}
\seealso{
See examples in \code{\link{dffit}}.
}
\author{
Danail Obreschkow
}
