% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfplotcov.R
\name{dfplotcov}
\alias{dfplotcov}
\title{Plot parameter covariance matrix}
\usage{
dfplotcov(bundle, p = NULL, n.points = 500, n.standarddev = 5,
  line.color = "blue", line.lwd = 2, point.color = "#bbbbff",
  point.cex = 0.1, text.size.labels = 1.1, text.size.numbers = 0.8,
  margins = c(4, 4, 0.5, 0.5), title = "")
}
\arguments{
\item{p}{Reference parameters to be plotted as crosses.}

\item{n.points}{Number of random points drawn from the multivariate Gaussian.}

\item{n.standarddev}{Number of standard-deviations shown to the left and right of the mean.}

\item{line.color}{Line color}

\item{line.lwd}{Line width}

\item{point.color}{Point color}

\item{point.cex}{Point size}

\item{text.size.labels}{Text size of parameter names}

\item{text.size.numbers}{Text size of parameter values}

\item{margins}{Margins (bottom,left,top,right)}

\item{title}{Title text}

\item{df}{List produced by \code{\link{dffit}}}
}
\description{
This function displays the parameter covariances  of the distribution function parameters (e.g. the mass function parameters) fitted using \code{\link{dffit}}.
}
\seealso{
See examples in \code{\link{dffit}}.
}
\author{
Danail Obreschkow
}
