% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfplotveff.R
\name{dfplotveff}
\alias{dfplotveff}
\title{Plot effective volume}
\usage{
dfplotveff(survey, xlab = "Observable x",
  ylab = expression("Effective volume"))
}
\arguments{
\item{xlab}{Label on x-axis (logarithmic mass or luminosity scale).}

\item{ylab}{Label on y-axis (volume scale).}

\item{df}{List produced by \code{\link{dffit}}}
}
\description{
This routine plots the effective survey volume function \code{Veff(x)} assumed when fitting a one-dimensional generative distribution function (GDF), such as a galaxy mass function. Note that this function \code{Veff(x)} is stored as \code{survey$selection$veff} when fitting a GDF using \code{survey=dffit(...)}.
}
\seealso{
See examples in \code{\link{dffit}}. To display \code{Veff(x)} of two-dimensional distribution functions use \code{\link{dfplotveff2}}.
}
\author{
Danail Obreschkow
}
