% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfplotveff2.R
\name{dfplotveff2}
\alias{dfplotveff2}
\title{Plot effective volume in 2D}
\usage{
dfplotveff2(df, xlab = "Observable x1", ylab = "Observable x2",
  zlab = "Veff")
}
\arguments{
\item{df}{List produced by \code{\link{dffit}}}

\item{xlab}{Label on x-axis (e.g. logarithm of mass).}

\item{ylab}{Label on y-axis (e.g. logarithm of size or angular momentum).}
}
\description{
This function plots the 2D function \code{Veff(x)}, where \code{x} is a 2-element vector.
}
\seealso{
See examples in \code{\link{dffit}}.
}
\author{
Danail Obreschkow
}
