% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfwrite.R
\name{dfwrite}
\alias{dfwrite}
\title{Write best fitting parameters}
\usage{
dfwrite(df)
}
\arguments{
\item{df}{List produced by \code{\link{dffit}}}
}
\description{
This function displays the best-fitting model parameters of a mass function previously fitted using \code{\link{dffit}}.
}
\examples{
data = mfdata()
df = dffit(data$x, data$selection, write.fit = FALSE)
dfwrite(df)

}
\author{
Danail Obreschkow
}
\seealso{
See examples in \code{\link{dffit}}.
}

