% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfwrite.R
\name{dfwrite}
\alias{dfwrite}
\title{Write best fitting parameters}
\usage{
dfwrite(survey)
}
\arguments{
\item{df}{List produced by \code{\link{dffit}}}
}
\description{
This function displays the best-fitting model parameters of a mass function previously fitted using \code{\link{dffit}}.
}
\examples{
dat = dfmockdata()
survey = dffit(dat$x, dat$veff, write.fit = FALSE)
dfwrite(survey)

}
\seealso{
See examples in \code{\link{dffit}}.
}
\author{
Danail Obreschkow
}
