% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfplot.R
\name{mfplot}
\alias{mfplot}
\title{Display fitted galaxy mass function}
\usage{
mfplot(df, xlab = expression("M [M"["sun"] * "]"), ylab = expression(phi ~
  "[Mpc"^-3 ~ "dex"^-1 ~ "]"), log = "xy", xpower10 = TRUE,
  show.data.points = TRUE, ...)
}
\description{
This function displays the galaxy mass function (MF) fitted using \code{\link{dffit}}, by calling \code{\link{dfplot}} with appropriate parameters.
}
\author{
Danail Obreschkow
}
\seealso{
See examples in \code{\link{dffit}}.
}

